backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
kombu==5.5.0
multidict==6.2.0
openai==1.67.0
prompt-toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.39
starlette==0.46.1
tqdm==4.67.1
twilio==9.5.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
